requirements.txt
streamlit
diffusers
torch
transformers
accelerate
pillow  # Do pracy z obrazami